Document store: use map instead of array, add save method
Before, it was not possible to override an existing entry in the
document store, which can cause issues if the same document is requested
and passed to `find()` twice (The document store would return the old
built document).

By switching to an object we can also make sure that each document is
only present once in the cache.