feat(jexl): add support for multiline expressions (#177)
This aims to make the jexl field in the form builder a bit more pleasant
to work with by allowing multiline expressions and rendering it with a
monospace font.

Multiline support in JEXL has already been added in the past, but was
(i suppose unintentionally) lost during the 2.0 release. See
https://github.com/TomFrost/Jexl/pull/54 for details.

pyJexl already supports multiline expressions.