feat(deps): drop support for ember 3.12
BREAKING CHANGE: The support for the former ember LTS version 3.12 was
dropped since we only support the latest LTS version which is now 3.16.

This change happened in 4eaecdb163a7cd6159d488c414dcf2b8afd98c27 where
we forgot to mark it as a breaking change which is why this commit is
empty.