fix(query): fix caluma query loop issue
This fixes the infinite loop bug that occured when `.fetch()` was
triggered by a `{{did-update}}` modifier. This also refactors the
queries internally to tasks to improve concurrency handling. Also, the
exposed properties `totalCount`, `hasNextPage` and `isLoading` are now
getters that make use of the last value of the fetch page task.