fix(distribution): fix which group can reopen an inquiry
Currently the `addressedGroups` of an inquiry can reopen it. Since they
are not responsible for the inquiry after completing it, this makes no
sense. Now, the `controllingGroups` of an inquiry can reopen them. Also,
this removes the complex logic to transition into the answer view since
the controlling group does not have to do anything after reopening the
inquiry.