fix(calc-questions): always calculate calculated answer if possible
We don't need to ensure that *every* dependency is non-hidden; indeed it
should be possible to evaluate an expression where parts of the dependencies
are not set. This is how the backend does it. Also, the `answer` transform
does have a default option, so we should easily be able to deal with this.