feat(form): leave valiation of format validators to backend
Format validators are currently evaluated in the backend and the
frontend. Since format validators now support custom validation without
regex, the frontend can't validate it anymore. As we didn't prevent the
save call anyway, the value was always validated in the backend but we
didn't do anything with the error message.

This commit changes the behaviour by not validating anything in the
frontend in favor of properly handling and displaying the error messaged
passed by the backend.